refactored scaling objects into it's own package
